Delete only plain URL messages

+++ src/controllers/instagram.ts
@@ -38,10 +38,12 @@ instagramController.on(
 					} catch (error) {
 						console.error('[IGC] Failed to respond with video', { directUrl })
 					}
-					try {
-						await ctx.deleteMessage()
-					} catch {
-						// ignore
+					if (text === url) {
+						try {
+							await ctx.deleteMessage()
+						} catch {
+							// ignore
+						}
 					}
 				}
 			} else {
